refactor(BonusBall): bonusNumber 원시값 포장 및 테스트 리팩토링
