+++ src/main/java/lotto/domain/Money.java
@@ -3,7 +3,8 @@ package lotto.domain;
 import lotto.exception.MoneyException;
 
 public class Money {
-    int money;
+    public static final int LOTTO_PRICE = 1000;
+    private final int money;
 
     public Money(int money) {
         validateMoneyValue(money);
@@ -11,12 +12,12 @@ public class Money {
     }
 
     private void validateMoneyValue(int money) {
-        if (money < 1000) {
+        if (money < LOTTO_PRICE) {
             throw new MoneyException();
         }
     }
 
-    public int getAffordableLottoTickets(int lottoPrice) {
-        return money / lottoPrice;
+    public int getAffordableLottoTickets() {
+        return money / LOTTO_PRICE;
     }
 }
