+++ src/test/java/lotto/controller/LottoStoreTest.java
@@ -2,6 +2,7 @@ package lotto.controller;
 
 import lotto.domain.*;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.CsvSource;
@@ -25,15 +26,6 @@ public class LottoStoreTest {
                 Arguments.of(new Lotto("2, 4, 6, 1, 7, 3"), 30000.00),
                 Arguments.of(new Lotto("1, 2, 3, 4, 5, 6"), 200000.00)
         );
-    }
-
-    @ParameterizedTest
-    @DisplayName("구입가능한 로또 매수 계산")
-    @CsvSource(value = {"14000,14", "10200,10", "500,0"})
-    void calculateAffordableLottoTicketsTest(int money, int expectedLottoTickets) {
-        LottoStore lottoStore = new LottoStore();
-        int calculatedLottoTickets = lottoStore.calculateAffordableLottoTickets(new Money(money));
-        assertThat(calculatedLottoTickets).isEqualTo(expectedLottoTickets);
     }
 
     @ParameterizedTest
